Added classifier.py
classifier.py is an ongoing file that should handle the problems of target_classifier_test.py, as that file seems incomplete. I wanted to start a new file instead of editing the original too much so that I could use the original as a baseline. Original file seemed to only test shape labeling and classifier.py should be close to classifying shape and letter. No code exists at all in the project to handle colors or orientation as it stands.